Fixed Log out issue which was causing crash.
Please make sure that when you do a fragment, everyone of them has a respective fragment class to be called from. What ended up happening was that after you completed the transaction via beginTransaction() call, it tried to replace the existing fragment with the current view. Since Log out fragment did not exist, it simply crashed.